project1/requirements.txt
numpy
plotly
ipywidgets==7.6.5
ipykernel
ipython
Pillow